+++ talpid-types/src/net/wireguard.rs
@@ -34,7 +34,7 @@ impl ConnectionConfig {
     pub fn get_endpoint(&self) -> Endpoint {
         Endpoint {
             address: self.peer.endpoint,
-            protocol: TransportProtocol::Udp,
+            protocol: self.peer.protocol,
         }
     }
 }
@@ -47,6 +47,14 @@ pub struct PeerConfig {
     pub allowed_ips: Vec<IpNetwork>,
     /// IP address of the WireGuard server.
     pub endpoint: SocketAddr,
+    /// Transport protocol. WireGuard only supports UDP directly.
+    /// If this is set to TCP, then traffic is proxied using [`udp_to_tcp::Udp2Tcp`].
+    #[serde(default = "default_peer_transport")]
+    pub protocol: TransportProtocol,
+}
+
+fn default_peer_transport() -> TransportProtocol {
+    TransportProtocol::Udp
 }
 
 #[derive(Clone, Eq, PartialEq, Deserialize, Serialize, Debug)]